Optimisations & fix for issue #3 using int buffers instead of short buffers

+++ src/com/monyetmabuk/rajawali/tutorials/ExampleParticleSystem.java
@@ -34,7 +34,7 @@ public class ExampleParticleSystem extends Particle {
 		float[] textureCoords = new float[numParticles * 2];
 		float[] normals = new float[numParticles * 3];
 		float[] colors = new float[numParticles * 4];
-		short[] indices = new short[numParticles];
+		int[] indices = new int[numParticles];
 		
 		int index = 0;
 		
@@ -62,7 +62,7 @@ public class ExampleParticleSystem extends Particle {
 			colors[i + 2] = i;
 			colors[i + 3] = i;
 			
-			indices[i] = (short)i;
+			indices[i] = i;
 		}
 		
 		mVelocityBuffer = ByteBuffer
